api: remove legacy settings.json editor, split phantom-config into schemas + storage
The old /ui/api/settings endpoint and the src/settings-editor module wrote
to ~/.claude/settings.json (Agent SDK internals). Operators who need those
knobs edit the file via SSH. The new /ui/api/phantom-config endpoint replaces
it end to end with zero overlap.

Also splits phantom-config.ts so each file stays under ~250 LOC:
- phantom-config-schemas.ts owns the Zod shapes and the derived types.
- phantom-config-storage.ts owns file IO, projection to the UI shape, the
  deep-merge patcher, and the write-plan builder.
- phantom-config.ts stays as the thin handler that orchestrates them.

No production consumer referenced the deleted surface.
